# Improved performance of csrf_input