@smithdc1 2022-01-18
If there is consensus this change is acceptable then there are a few similar cases we should consider. The most obvious being `CsrfTokenNode`