@kezabelle 2022-01-19
Computers forever proving they're hard; I see an improvement, as you did, though the magnitude is different:
```
In [9]: %timeit csrf_input_new(r)
74.5 µs ± 516 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [10]: %timeit csrf_input(r)
80 µs ± 755 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

The CSRF stuff is complex enough and defensive enough that I can't immediately tell if there's any implications from assuming it's trusted -- it _looks like_ in this case it is maybe trustable, because it'll either give back a masked value based on `ascii_letters + digits` or I think raise an exception or reject the incoming value (due to length or invalid characters by regex search) elsewhere. 
I'm less certain about `CsrfTokenNode` because that resolves the value out of the context, assuming it has come from the context processor ... but what if it didn't? I dunno :)